Use ~/.pyenv as the pyenv directory
Using /workspace/.pyenv resulted in a bunch of complications related to
the Gitpod envirnoment.

This means that mew Python versions installed in the workspace will not
persist (following a restart or from the prebuild stage), but typically
the version we need will already be installed by the Dockerfile so it
won't matter much in practice.